LPS-146033 search-experiences-web: Create dedicated folder for hooks separate from utils
